refactor: separando lógica com custom hooks
