+++ mobile/src/pages/NCut/index.tsx
@@ -1,142 +1,55 @@
-import React, {useState, useEffect, useContext} from 'react';
-import {Alert, ScrollView, StyleSheet, View} from 'react-native';
+/* eslint-disable react-hooks/exhaustive-deps */
+import React, {
+  useState,
+  useEffect,
+  useContext,
+  Dispatch,
+  SetStateAction,
+} from 'react';
+import {ScrollView, StyleSheet, View} from 'react-native';
 import {NavigationProp} from '@react-navigation/native';
 import {useRoute} from '@react-navigation/native';
 
 import UserContext from '../../context/GlobalContext';
-import strapi from '../../config/strapi';
 import NCutSlider from 'components/NCutSlider';
 import AppText from 'components/ui/AppText';
 import Button from 'components/ui/Button';
+import useNCut from '../../hooks/useNCut';
 
 type NCutType = {
   navigation: NavigationProp<any, any>;
 };
 
-async function getAreas(id: number) {
-  return strapi
-    .findOne('propostas', id, {
-      fields: ['areas', 'valor'],
-    })
-    .then((data: any) => {
-      return data.data;
-    })
-    .catch(error => {
-      console.log(error);
-    });
-}
+type slideType = {
+  id: number;
+  key: string;
+  area: string;
+  valorInicial: number;
+  valorMaximo: number;
+  setValores: Dispatch<SetStateAction<{[key: number]: number}>>;
+  valorMinimo: number;
+};
 
-async function submitVote(
-  userId: number,
-  proposta: number,
-  voto: {[key: number]: number},
-) {
-  return strapi
-    .create('n-cuts', {
-      proposta: {
-        connect: [proposta],
-      },
-      usuario: {
-        connect: [userId],
-      },
-      voto: voto,
-    })
-    .then((data: any) => {
-      return data.data;
-    })
-    .catch(error => {
-      console.log(error);
-    });
-}
-
-function getFormattedValue(value: number) {
-  let finalCost = String(value.toFixed(2)).replace('.', ',');
-  finalCost = finalCost.replace(/\d{1,3}(?=(\d{3})+(?!\d))/g, '$&.');
-  return finalCost;
-}
-
-function votacaoStrings(areas: String[], valores: {[key: number]: number}) {
-  let string: string = '';
-  let soma = 0;
-  for (let i = 0; i < areas.length; i++) {
-    string += areas[i] + ': R$ ' + getFormattedValue(valores[i]) + '\n';
-    soma += valores[i];
-  }
-  string += '\nValor total: R$ ' + getFormattedValue(soma);
-  return string;
+function valorColor(soma: number, valorMaximo: number) {
+  return {
+    textAlign: 'center',
+    fontSize: 20,
+    fontWeight: 'bold',
+    color: soma > valorMaximo ? 'red' : 'black',
+  };
 }
 
 export default function NCut({navigation}: NCutType) {
   const [valores, setValores] = useState<{[key: number]: number}>({});
-  const [sliders, setSliders] = useState<any[]>([]);
+  const [sliders, setSliders] = useState<slideType[]>([]);
   const [valorAlocado, setValorAlocado] = useState(0);
   const [valorMaximo, setValorMaximo] = useState(0);
-  const [areas, setAreas] = useState<String[]>([]);
+  const [areas, setAreas] = useState<string[]>([]);
   const {userId} = useContext(UserContext);
+  const {getAreas, getFormattedValue, onSubmit} = useNCut();
 
   const route: any = useRoute();
-
   let id = parseInt(route.params.id, 10);
-
-  function onSubmit() {
-    if (valorAlocado > valorMaximo) {
-      Alert.alert(
-        'Valor inválido!',
-        'O valor que você alocou é maior que o orçamento total.',
-      );
-    } else {
-      let valoresAlert: string = votacaoStrings(areas, valores);
-      Alert.alert('Confira os valores de sua votação: ', valoresAlert, [
-        {text: 'Cancelar', style: 'cancel'},
-        {
-          text: 'Enviar voto',
-          onPress: () => {
-            submitVote(userId, id, valores)
-              .then(data => {
-                if (data !== undefined) {
-                  if (Object.keys(data).length > 0) {
-                    Alert.alert(
-                      'Seu voto foi registrado!',
-                      'Obrigado por participar desta votação.',
-                      [{text: 'OK', onPress: () => navigation.goBack()}],
-                    );
-                  }
-                } else {
-                  Alert.alert(
-                    'Algo deu errado!',
-                    'Infelizmente, não conseguimos registrar o seu voto no nosso banco de dados. Por favor, tente novamente! Se mesmo assim você não conseguir, entre em contato com os responsáveis pelo app.',
-                  );
-                }
-              })
-              .catch(() => {
-                Alert.alert(
-                  'Algo deu errado!',
-                  'Infelizmente, não conseguimos registrar o seu voto no nosso banco de dados. Por favor, tente novamente! Se mesmo assim você não conseguir, entre em contato com os responsáveis pelo app.',
-                );
-              });
-          },
-        },
-      ]);
-    }
-  }
-
-  function valorColor(soma: number) {
-    if (soma > valorMaximo) {
-      return {
-        textAlign: 'center',
-        fontSize: 20,
-        fontWeight: 'bold',
-        color: 'red',
-      };
-    } else {
-      return {
-        textAlign: 'center',
-        fontSize: 20,
-        fontWeight: 'bold',
-        color: 'black',
-      };
-    }
-  }
 
   useEffect(() => {
     let soma = 0;
@@ -149,10 +62,9 @@ export default function NCut({navigation}: NCutType) {
   useEffect(() => {
     getAreas(id)
       .then(data => {
-        let slidersLoad: Array<any> = [];
+        let slidersLoad: Array<slideType> = [];
         let areasAtt: {[key: string]: string} = data.attributes.areas;
         let valor = data.attributes.valor;
-
         setValorAlocado(valor);
         setValorMaximo(valor);
 
@@ -163,17 +75,15 @@ export default function NCut({navigation}: NCutType) {
         for (const k in areasAtt) {
           valoresIniciais[parseInt(k, 10)] = Math.floor(valor / size);
           areasStrings.push(areasAtt[k]);
-          slidersLoad.push(
-            <NCutSlider
-              id={parseInt(k, 10)}
-              key={k}
-              area={areasAtt[k]}
-              valorInicial={Math.floor(valor / size)}
-              valorMaximo={valor}
-              setValores={setValores}
-              valorMinimo={0}
-            />,
-          );
+          slidersLoad.push({
+            id: parseInt(k, 10),
+            key: k,
+            area: areasAtt[k],
+            valorInicial: Math.floor(valor / size),
+            valorMaximo: valor,
+            setValores: setValores,
+            valorMinimo: 0,
+          });
         }
         setSliders(slidersLoad);
         setValores(valoresIniciais);
@@ -193,14 +103,36 @@ export default function NCut({navigation}: NCutType) {
             Utilize os seletores abaixo ou as caixas de texto para especificar
             os valores que você deseja que sejam alocados para cada área.
           </AppText>
-          {sliders}
+          {sliders.map(slider => (
+            <NCutSlider
+              key={slider.key}
+              id={slider.id}
+              area={slider.area}
+              valorInicial={slider.valorInicial}
+              setValores={slider.setValores}
+              valorMaximo={slider.valorMaximo}
+              valorMinimo={slider.valorMinimo}
+            />
+          ))}
         </View>
-        <Button style={styles.buttonStyle} clickFn={onSubmit}>
+        <Button
+          style={styles.buttonStyle}
+          clickFn={() =>
+            onSubmit(
+              id,
+              userId,
+              areas,
+              valorAlocado,
+              valorMaximo,
+              valores,
+              navigation,
+            )
+          }>
           <AppText style={styles.buttonText}>Enviar Voto</AppText>
         </Button>
       </ScrollView>
       <View style={styles.navbar}>
-        <AppText style={valorColor(valorAlocado)}>
+        <AppText style={valorColor(valorAlocado, valorMaximo)}>
           Valor Alocado: R$ {getFormattedValue(valorAlocado)}
         </AppText>
         <AppText style={styles.smallDescription}>
